Fix CLI connection to watchman-to-sql command
